+++ wilson_flow/setup.c
@@ -76,7 +76,7 @@ initial_set()
   nt=par_buf.nt;
   
   number_of_nodes = numnodes();
-  volume=nx*ny*nz*nt;
+  volume=(size_t)nx*ny*nz*nt;
   
   return prompt;
 }
